fix: use session for pdpa, move authOptions to lib
