+++ src/app/(auth)/profile/page.tsx
@@ -1,7 +1,7 @@
 import { prisma } from "@/lib/prisma";
 import { getServerSession } from "next-auth";
 import { redirect } from "next/navigation";
-import { authOptions } from "../../api/auth/[...nextauth]/auth";
+import { authOptions } from "../../../lib/auth";
 import { ProfileForm } from "./ProfileForm";
 
 async function getUserProfile(userId: string) {
