+++ src/app/api/admin/verify-seller/route.ts
@@ -1,7 +1,7 @@
 import { prisma } from "@/lib/prisma";
 import { getServerSession } from "next-auth/next";
 import { NextResponse } from "next/server";
-import { authOptions } from "../../auth/[...nextauth]/auth";
+import { authOptions } from "../../../../lib/auth";
 
 export async function POST(req: Request) {
   const session = await getServerSession(authOptions);
