+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,5 +1,5 @@
 import NextAuth from "next-auth";
-import { authOptions } from "./auth";
+import { authOptions } from "../../../../lib/auth";
 
 const handler = NextAuth(authOptions);
 export { handler as GET, handler as POST };
