+++ src/app/api/auth/seller-registration/route.ts
@@ -2,7 +2,7 @@ import { prisma } from "@/lib/prisma";
 import { put } from "@vercel/blob";
 import { getServerSession } from "next-auth";
 import { NextResponse } from "next/server";
-import { authOptions } from "../[...nextauth]/auth";
+import { authOptions } from "../../../../lib/auth";
 
 export async function POST(request: Request) {
   try {
